app/requirements.txt
fastapi
uvicorn
openai
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
pydantic
tiktoken